Endpoints for Instituion are refactored
