+++ backend/TimeTile/TimeTile.API/Endpoints.cs
@@ -8,9 +8,9 @@ using TimeTile.API.ClassroomTypes.Endpoints.GetById;
 using TimeTile.API.Common.Api;
 using TimeTile.API.Common.Api.Extensions;
 using TimeTile.API.Common.Api.Filters;
-using TimeTile.API.Institutions.Endpoints.CreateInstitution;
-using TimeTile.API.Institutions.Endpoints.GetInstitutionById;
-using TimeTile.API.Institutions.Endpoints.GetInstitutions;
+using TimeTile.API.Institutions.Endpoints.Create;
+using TimeTile.API.Institutions.Endpoints.GetById;
+using TimeTile.API.Institutions.Endpoints.Get;
 using TimeTile.API.LessonStatuses.Endpoints.Create;
 using TimeTile.API.LessonStatuses.Endpoints.Get;
 using TimeTile.API.LessonStatuses.Endpoints.GetById;
