+++ backend/TimeTile/TimeTile.API/Institutions/Endpoints/Get/RequestValidator.cs
@@ -1,7 +1,7 @@
 ﻿using TimeTile.API.Common.Api.Extensions;
 using TimeTile.API.Common.Api.Pagination.PagedRequest;
 
-namespace TimeTile.API.Institutions.Endpoints.GetInstitutions;
+namespace TimeTile.API.Institutions.Endpoints.Get;
 
 public class RequestValidator : PagedRequestValidator<GetInstitutionsEndpoint.Request>
 {
